# Conflicts:
#	src/main/java/com/ipi/mesi_backend_rpg/controller/ModuleController.java
#	src/main/java/com/ipi/mesi_backend_rpg/model/Module.java
#	src/main/java/com/ipi/mesi_backend_rpg/model/ModuleVersion.java
#	src/main/java/com/ipi/mesi_backend_rpg/repository/ModuleRepository.java
#	src/main/java/com/ipi/mesi_backend_rpg/service/ModuleService.java

+++ src/main/java/com/ipi/mesi_backend_rpg/dto/TagDTO.java
@@ -4,10 +4,9 @@ import com.ipi.mesi_backend_rpg.model.Module;
 
 import java.util.List;
 
-public record TagDTO (
-        Integer id,
-        String name,
-        List<Module> modules
-){
+public record TagDTO(
+                Long id,
+                String name,
+                List<Module> modules) {
 
 }
